# perf: Optimized SQL queries for indexing logic