--- a/mobile/src/modules/scanning/helpers/artwork.ts
+++ b/mobile/src/modules/scanning/helpers/artwork.ts
@@ -1,8 +1,9 @@
 import { getArtwork } from "@missingcore/react-native-metadata-retriever";
-import { eq, isNotNull } from "drizzle-orm";
+import { eq, inArray, isNotNull, or } from "drizzle-orm";
 import * as FileSystem from "expo-file-system";
 
 import { db } from "@/db";
+import type { TrackWithAlbum } from "@/db/schema";
 import { albums, artists, playlists, tracks } from "@/db/schema";
 
 import { getAlbums, updateAlbum } from "@/api/album";
@@ -19,10 +20,30 @@ import { batch } from "@/utils/promise";
 export async function findAndSaveArtwork() {
   const stopwatch = new Stopwatch();
 
-  const uncheckedTracks = await getTracks([eq(tracks.fetchedArt, false)]);
+  // Ensure we don't unnecessarily seach for artwork.
   const albumsWithCovers = new Set(
     (await getAlbums([isNotNull(albums.artwork)])).map(({ id }) => id),
   );
+  await db
+    .update(tracks)
+    .set({ fetchedArt: true })
+    .where(
+      or(
+        inArray(tracks.albumId, [...albumsWithCovers]),
+        isNotNull(tracks.artwork),
+      ),
+    );
+
+  const uncheckedTracks = await getTracks([eq(tracks.fetchedArt, false)]);
+  // Sort tracks to optimize SQL queries.
+  const singles: TrackWithAlbum[] = [];
+  const albumTracks: Record<string, TrackWithAlbum[]> = {};
+  uncheckedTracks.forEach((t) => {
+    const key = t.albumId;
+    if (key === null) singles.push(t);
+    else if (Object.hasOwn(albumTracks, key)) albumTracks[key]!.push(t);
+    else albumTracks[key] = [t];
+  });
 
   // Initiate the image saving phase.
   onboardingStore.setState({
@@ -34,46 +55,77 @@ export async function findAndSaveArtwork() {
 
   let newArtworkCount = 0;
   let checkedFiles = 0;
+  let prevRemainder = 0;
 
-  for (const { id, albumId, uri, name, artwork } of uncheckedTracks) {
-    // Indicate we attempted to find artwork for a track. Do this before
-    // physically attempting to save the artwork in case an OOM error occurs,
-    // in which the app will essentially become "bricked".
-    await updateTrack(id, { fetchedArt: true });
-    checkedFiles++;
+  // Get artwork for albums.
+  for (const [albumId, values] of Object.entries(albumTracks)) {
+    await saveSinglesArtwork(
+      values,
+      async ({ artworkUri }) => {
+        await updateAlbum(albumId, { artwork: artworkUri });
+        albumsWithCovers.add(albumId);
+        newArtworkCount++;
+      },
+      { endEarly: true },
+    );
     // Prevent excessive `setState` on Zustand store which may cause an
     // "Warning: Maximum update depth exceeded.".
-    if (checkedFiles % 25 === 0) {
+    prevRemainder = checkedFiles % 25;
+    checkedFiles += values.length;
+    if (checkedFiles % 25 < prevRemainder) {
       onboardingStore.setState({ checked: checkedFiles });
     }
-
-    // Make sure the track doesn't have `artwork` and either be unassociated
-    // with an album or its album doesn't have `artwork`.
-    if (!artwork && (!albumId || !albumsWithCovers.has(albumId))) {
-      try {
-        const base64Artwork = await getArtwork(uri);
-        if (base64Artwork) {
-          const artwork = await saveImage(base64Artwork);
-          if (albumId) {
-            await updateAlbum(albumId, { artwork });
-            albumsWithCovers.add(albumId);
-          } else {
-            await updateTrack(id, { artwork });
-          }
-          newArtworkCount++;
-          onboardingStore.setState((prev) => ({ found: prev.found + 1 }));
-        }
-      } catch (err) {
-        // In case we fail to save an image due to having an invalid base64 string.
-        console.log(`[Error] Failed to get or save image for "${name}".`);
-      }
-    }
   }
+
+  // Get artwork for tracks.
+  await saveSinglesArtwork(
+    singles,
+    async ({ artworkUri, trackId }) => {
+      await updateTrack(trackId, { artwork: artworkUri });
+      newArtworkCount++;
+    },
+    {
+      onEndIteration: () => {
+        // Prevent excessive `setState` on Zustand store which may cause an
+        // "Warning: Maximum update depth exceeded.".
+        checkedFiles += 1;
+        if (checkedFiles % 25 === 0) {
+          onboardingStore.setState({ checked: checkedFiles });
+        }
+      },
+    },
+  );
   console.log(
     `Finished saving ${newArtworkCount} new cover images in ${stopwatch.lapTime()}.`,
   );
 
   if (newArtworkCount > 0) clearAllQueries();
+}
+
+async function saveSinglesArtwork(
+  singles: TrackWithAlbum[],
+  onSave: (info: { artworkUri: string; trackId: string }) => Promise<void>,
+  options?: { endEarly?: boolean; onEndIteration?: () => void },
+) {
+  for (const { id: trackId, uri, name } of singles) {
+    // Indicate we attempted to find artwork for a track. Do this before
+    // physically attempting to save the artwork in case an OOM error occurs,
+    // in which the app will essentially become "bricked".
+    await updateTrack(trackId, { fetchedArt: true });
+    try {
+      const base64Artwork = await getArtwork(uri);
+      if (base64Artwork) {
+        const artworkUri = await saveImage(base64Artwork);
+        await onSave({ artworkUri, trackId });
+        onboardingStore.setState((prev) => ({ found: prev.found + 1 }));
+        if (options?.endEarly) return;
+      }
+    } catch (err) {
+      // In case we fail to save an image due to having an invalid base64 string.
+      console.log(`[Error] Failed to get or save image for "${name}".`);
+    }
+    if (options?.onEndIteration) options.onEndIteration();
+  }
 }
 //#endregion