# Why

<!--
Please describe the motivation for this PR, and link to relevant GitHub issues, discussions, or feature requests.
-->

Around a week ago, we applied some SQL optimizations to the import feature where we did bulk insertion of values instead of inserting an entry one-by-one in a loop. From that experiment, we saw a decent improvement in the time it takes from reading the file to the saving those values in the database (we went from ~2.7s to ~1.3s). One of the places this idea can also be applied to is the indexing logic, or the logic which is run to find and save the metadata of the various tracks found on the device.

Some observations from applying these optimizations include:
- A ~2.4s reduction (~9.6s to ~7.2s) in the time it takes to save the 25 artwork/covers found from the 238 tracks available in **development** mode.
- We saw no impactful improvement from optimization tests for the main indexing (getting the metadata excluding the artwork and saving those values into the database).
  - For my 238 tracks, in **development** mode, it takes around ~11s to finish this task. Around ~10s is used exclusively for getting the metadata of all the tracks. So essentially, we have it pretty optimized already.
- As a test, since I never actually done this, I timed how long it takes for a fresh install in production to reach the home screen of the app (from the time we click the permissions to where we see the home screen).
  - It took ~19s in **development** mode and ~16.5s in **production** mode.

**Other Changes:**
- Removed the batching logic used in the function for removing albums and artists with no content.
- Fixed the type signature for `onBatchComplete` in our `batch()` function.

# How

<!--
How did you build this feature or fix this bug and why?
-->

# Test Plan

<!--
Please describe how you tested this change and how a reviewer could reproduce your test, especially if this PR does not include automated tests! If possible, please also provide terminal output and/or screenshots demonstrating your test/reproduction.
-->

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
-->

- [ ] Documentation is up to date to reflect these changes (ie: `CHANGELOG.md` & `README.md`).
- [ ] Add new dependencies into `THIRD_PARTY.md`.
- [x] This diff will work correctly for `pnpm android:prod`.